Add tests for namespacing and twiml constructors
